fix(api): Return all relevant subscriber fields as test data
Return all subscriber fields as the test data.to field based on the variables used in the workflow and the steps that the workflow includes.

That is, an SMS step requires a phone, an Email step an email and any other field should be present in the To test data object if it's used as a variable in the content of the workflow steps.